Add a PUT /:id endpoint to our route file
In this commit, I added a PUT /:id Endpoint to our api router file. this endpoint
is used to update the details about a particular car. This endpoint uses two midd-
le ware namely validateCarParams() and validateCarId() to perform preliminary check
on the car details to be inserted and also check if the car id exists respectively